src/main/java/practice/leecode/LinkedTable/RemoveNthFromEnd_19.java
package practice.leecode.LinkedTable;

import practice.utils.ListNode;

public class RemoveNthFromEnd_19 {

    public ListNode removeNthFromEnd(ListNode head, int n) {
        ListNode prev = new ListNode(0);
        prev.next = head;
        ListNode slow = prev;
        ListNode fast = prev;
        for (int i = 0; i <= n; i++) {
            fast = fast.next;
        }
        while (fast != null) {
            slow = slow.next;
            fast = fast.next;
        }
        slow.next = slow.next.next;
        return prev.next;
    }

}
